Add image to process description
